+++ lib/zakuro/calculation/monthly/internal/meta/all_solar_term.rb
