+++ lib/zakuro/calculation/monthly/internal/operated_solar_term.rb
