+++ lib/zakuro/calculation/monthly/internal/part/const.rb
