+++ lib/zakuro/calculation/monthly/internal/part/month_label.rb
